Add terminal encoding option
By default, VTE initializes a UTF8 terminal (which is specifically instantiated here as such via the
VTE_ENCODING constant). However, not all applications are UTF8-aware, and some rely on hacks or
alternative encodings in order to display certain characters (*cough* *cough* Nethack).

This commit adds the required hooks and documentation for allowing us to change the terminal
encoding scheme to something other than UTF8, in an effort to cope with these broken applications.

The option character `t` was the closest thing to 'terminal encoding' I could find.